# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gestion_Produit API",
    "version": "1.0.0",
    "description": "CRUD produits, catégories & fournisseurs."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Dev local"
    }
  ],
  "components": {
    "schemas": {
      "Categorie": {
        "type": "object",
        "required": [
          "libelle"
        ],
        "properties": {
          "libelle": {
            "type": "string",
            "example": "Charcuterie"
          }
        }
      },
...
      "Fournisseur": {
        "type": "object",
        "required": [
          "prenom",
          "nom",
          "adresse",
          "email",
          "telephone"
        ],
        "properties": {
          "prenom": {
            "type": "string",
            "example": "Fatima"
          },
          "nom": {
            "type": "string",
            "example": "Niang"
          },
          "adresse": {
            "type": "string",
            "example": "Dakar"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "niang@gmail.com"
          },
          "telephone": {
            "type": "string",
            "pattern": "^\\+221(77|78|76)[0-9]{7}$",
            "example": "+221771234567"
          }
        }
      },
...
      "Produit": {
        "type": "object",
        "required": [
          "libelle",
          "qte",
          "prix",
          "categoryId",
          "fournisseurId"
        ],
        "properties": {
          "libelle": {
            "type": "string",
            "example": "Tomate"
          },
          "qte": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "example": 5
          },
          "prix": {
            "type": "number",
            "example": 1500
          },
          "categoryId": {
            "type": "integer",
            "example": 1
          },
          "fournisseurId": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
    },
...
  },
  "paths": {
    "/categories": {
      "get": {
        "summary": "Récupérer toutes les catégories",
        "responses": {
          "200": {
            "description": "Liste des catégories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Ajouter une catégorie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Catégorie créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorie"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
...
    "/fournisseurs": {
      "get": {
        "summary": "Récupérer tous les fournisseurs",
        "responses": {
          "200": {
            "description": "Liste des fournisseurs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fournisseur"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Ajouter un fournisseur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Fournisseur"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fournisseur créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fournisseur"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
...
    "/produits": {
      "get": {
        "summary": "Récupérer tous les produits",
        "responses": {
          "200": {
            "description": "Liste des produits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Produit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Ajouter un produit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Produit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produit"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
...
  }
}